requirements.txt
pandas
numpy
matplotlib
ipykernel
nltk
tqdm
wordcloud
scikit-learn
gdown
pyyaml